+++ frontend/src/pages/homePage/components/graphMenu/GraphMenu.style.ts
