+++ frontend/src/pages/homePage/components/graphMenuTab/GraphMenuTab.style.ts
