+++ frontend/src/pages/homePage/components/headLine/HeadLine.style.ts
