+++ frontend/src/pages/homePage/components/mainPageGraph/MainPageGraph.style.ts
