+++ frontend/src/pages/homePage/components/mainPageGraph/MainPageGraph.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import GraphMenu from '../graphMenu/GraphMenu';
-import S from './MainPageGraph.styles';
+import S from './MainPageGraph.style';
 
 const MainPageGraph = () => {
   const [tabIndex, setTabIndex] = useState(0);
