+++ frontend/src/pages/homePage/components/workCell/WorkCell.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import DragIcon from '@/assets/icons/flat/IC24Drag.svg?react';
 import S from './WorkCell.style';
 import { ColorPrimary } from '@/styles/colors';
-import type { WorkCellType, WorkCellStatus } from '@/types/workCell.types';
+import type { WorkCellType, WorkCellStatus } from '@/types/workCell.type';
 
 interface WorkCellProps {
   type: WorkCellType;
