+++ frontend/src/pages/myFarmPage/components/cropInfoCard/CropInfoCard.style.ts
