+++ frontend/src/types/workCell.type.ts
